:zap: [improve/coin-limit] 웹툰 탑승/하차 코인 보상 횟수 제한 코드 리팩토링 & 웹툰 탑승/하차 시 코인 히스토리 저장 내용 json으로 변경

+++ src/main/java/kr/co/antoon/coin/AntCoinClient.java
@@ -43,5 +43,5 @@ public interface AntCoinClient {
     /**
      * @apiNote limit of coin reward for join webtoon
      **/
-    int rewardLimit = 10;
+    Long rewardLimit = Long.valueOf(10);
 }
