+++ src/main/java/kr/co/antoon/coin/infrastructure/AntCoinHistoryRepository.java
@@ -13,5 +13,5 @@ public interface AntCoinHistoryRepository extends JpaRepository<AntCoinHistory, 
     boolean existsByRemittanceTypeAndUserIdAndReasonAndCreatedAtAfter(RemittanceType remittanceType, Long userId, String reason, LocalDateTime now);
 
     CoinHistory getAntCoinHistoryByUserId(Long userId);
-    int countByUserIdAndCreatedAtAfter(Long userId, LocalDateTime now);
+    Long countByUserIdAndCreatedAtAfter(Long userId, LocalDateTime now);
 }
